+++ packages/web/src/pages/Main/ts/deletePersonLogic.ts
@@ -1,5 +1,13 @@
-import { removeDisabledAttributeByID, setDisabledAttributeByID } from '../../utils/ts/utils';
+import { removeDisabledAttributeByID, setDisabledAttributeByID } from '../../../utils/ts/utils';
 import { closedModal } from './modal';
+
+export function filterObjInState(stateArr, id) {
+  return stateArr.filter((el) => {
+    if (el.id !== Number(id)) {
+      return el;
+    }
+  });
+}
 
 export function deleteRow(state): boolean {
   setDisabledAttributeByID('saveDelete');
@@ -14,25 +22,18 @@ export function deleteRow(state): boolean {
       }
 
       if (response.status === 200) {
-        state.currentData = state.currentData.filter((el) => {
-          if (el.id !== Number(state.currentSelectedNode.id)) {
-            return el;
-          }
-        });
+        state.currentData = filterObjInState(state.currentData, state.currentSelectedNode.id);
 
         if (state.currentSortedData !== null) {
-          state.currentSortedData = state.currentSortedData.filter((el) => {
-            if (el.id !== Number(state.currentSelectedNode.id)) {
-              return el;
-            }
-          });
+          state.currentSortedData = filterObjInState(state.currentSortedData, state.currentSelectedId);
         }
 
         state.currentSelectedNode.parentNode.removeChild(state.currentSelectedNode);
-        setDisabledAttributeByID('buttonDelete');
-        setDisabledAttributeByID('buttonUpdate');
         state.currentSelectedNode = null;
         state.currentSelectedId = null;
+        state.currentSelectedObj = null;
+        setDisabledAttributeByID('buttonDelete');
+        setDisabledAttributeByID('buttonUpdate');
         removeDisabledAttributeByID('saveDelete');
         closedModal('deleteModal');
         return true;
